+++ DVD.cs
@@ -10,7 +10,7 @@ namespace csharp_biblioteca_db
     {
         public int Durata { get; set; }
 
-        public DVD(int Codice, string Titolo, string Settore,string Scaffale, int Durata) : base(Codice, Titolo, Settore, Scaffale)
+        public DVD(long Codice, string Titolo, string Settore,string Scaffale, int Durata) : base(Codice, Titolo, Settore, Scaffale)
         {
             this.Durata = Durata;
